requirements.txt
Flask==2.3.3
pymongo==4.4.0
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
Werkzeug==2.3.7